Updated the gitignore
Added a client app, this is going to be used to download the new client. So when you navigate to /client in the server you are able to download the latest version

Removed unused files

Moved api url related stuff to the api application, just include it to the main urls.py file in the cmt dir

After some debugging, found some working code. Shall be adding all the models to the api. Only then I can make references between the models

Added more Serializers, Filters and Views. Also adding views witouth the plural name, due to the fact of auto-detect feature of the django-restframework. This way the code is more clean

Added more and more models to the api

finalized apilizing the models, now testing if everything is working

Just enable the new api for this, but we still call this v1

Added dynamic model in cmt_client generation

Reading, deleting and parsing templates is working in this API. Still need fix the api for creating, updating

Renamed old client to old_client and remooved unused file

Moved all server code to root of project